Versão 1.1.151.182

+++ web/WEB-INF/metadata/migrate/20150928083434_add_column_modulosPermitidoSomatorio_parametrosDoSistema.sql
@@ -1,1 +1,1 @@
-ALTER TABLE parametrosdosistema ADD COLUMN modulosPermitidosSomatorio bigint DEFAULT 63;--.go
+ALTER TABLE parametrosdosistema ADD COLUMN modulosPermitidosSomatorio smallint DEFAULT 63;--.go
